Add files via upload
To access 'dashboard', open anaconda prompt and type jupyter lab and run it. A new window for jupyter lab will open, then find this 'dashboard' file in downloads. 
Give the file path of the csv/excel/pickle file of the dataset used for analysis and read it using pd.read_csv/excel('file path') or use pickle. 
Run the codes available in the 'dashboard', you will get the interactive plots for various columns available in the dataframe.